scripts/requirements.txt
pyannote.audio
openai-whisper
torch
torchaudio
